pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zee-api"
version = "0.1.3"
authors = [
    { name = "Lucas Mello dos Santos", email = "contact@lmello.dev" }
]
description = "Zee-API framework implemented on top of FastAPI."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E"]
dependencies = [
    "fastapi==0.118.0",
    "pydantic-settings==2.11.0",
    "PyYAML>=6.0.3"
]